fixed overflow of sidebar and restricted it till chatbody
