+++ packages/react/src/views/ChatLayout/ChatLayout.styles.js
@@ -1,0 +1,10 @@
+import { css } from '@emotion/react';
+
+const styles = {
+  main: css`
+    overflow: hidden;
+    display: flex;
+  `,
+};
+
+export default styles;
