+++ packages/react/src/views/ChatLayout/index.js
@@ -1,0 +1,1 @@
+export { default as ChatLayout } from './ChatLayout';
